code/solution.no2/server/requirements.txt
google-genai==1.31.0
fastapi==0.116.1
uvicorn[standard]